tests/wpt/web-platform-tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.put.wrongtype.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.imageData.put.wrongtype</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.imageData.put.wrongtype</h1>
<p class="desc">putImageData() does not accept non-ImageData objects</p>


<script>
var t = async_test("putImageData() does not accept non-ImageData objects");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

var imgdata = { width: 1, height: 1, data: [255, 0, 0, 255] };
assert_throws_js(TypeError, function() { ctx.putImageData(imgdata, 0, 0); });
assert_throws_js(TypeError, function() { ctx.putImageData("cheese", 0, 0); });
assert_throws_js(TypeError, function() { ctx.putImageData(42, 0, 0); });
t.done();

});
</script>
